Update migration & seeder

+++ app/Database/Seeds/KamartarifSeeder.php
@@ -9,13 +9,14 @@ class KamartarifSeeder extends Seeder
 {
     public function run()
     {
-        $r = (new KamartarifModel())->insert([
+        $r = (int)(new KamartarifModel())->insert([
             'kamartipe_id' => 1,
             'tarif' => 2000000,
             'tgl_mulai' => '2022-11-02',
             'tgl_selesai' => '2022-11-05',
             'tipetarif_id' => 1,
         ]);
+
         echo "hasil insert $r";
     }
 }
